% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAttributes.R
\name{addAttributes}
\alias{addAttributes}
\title{Add attributes to \code{puml$code}}
\usage{
addAttributes(x, nm, puml)
}
\arguments{
\item{x}{the object}

\item{nm}{the name}

\item{puml}{the plantuml object}
}
\value{
puml
}
\description{
Internal Only
}
