% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addInfo}
\alias{addInfo}
\title{Add info to \code{puml$code}}
\usage{
addInfo(x, nm, puml)
}
\arguments{
\item{x}{the object}

\item{nm}{the name}

\item{puml}{the plantuml object}
}
\value{
puml
}
\description{
Internal Only
}
