% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.plantuml.complex.R
\name{as.plantuml.complex}
\alias{as.plantuml.complex}
\title{Convert a character to a \code{plantuml} object}
\usage{
\method{as.plantuml}{complex}(x, complete = FALSE, nm = NULL, ...)
}
\arguments{
\item{x}{character sting containing plantuml code.}

\item{complete}{if \code{TRUE}, enclose the plantuml statements in \code{"@startuml"} and \code{"@enduml"}}

\item{nm}{normally not for user - TODO I have to check... embarassing}

\item{...}{additional arguments - not used here}
}
\value{
object of class \code{plantuml} which can be plotted.
}
\description{
Convert a \code{character} to a \code{plantuml} object.
This can be plotted.
}
\examples{
\dontrun{
 x <- '
 @startuml --> "First Activity" -->[You can put also labels] "Second Activity"
 -->
 @enduml
 '
 x <- plantuml( x )
 plot( x ) }

}
