% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_yaml_json.R
\name{diff_yaml_json}
\alias{diff_yaml_json}
\title{Create diff between two lists to be used in the `plantuml_yaml()` and
`plantuml_json()` as preamble to highlight the differences.}
\usage{
diff_yaml_json(x, y)
}
\arguments{
\item{x}{list basde on yaml or json file to be compared. See the details zection for further info.}

\item{y}{list based on yaml or json file to be compared. See the details zection for further info.}
}
\value{
`character` vector to be used as preamble for the `plot_yaml()` and
  `plantuml_json()` functions (or `plantuml_file()` in case of a f=json or
  yaml file..
}
\description{
This function requires that the names of `x` and `y` are identical and in the
same order. Any deviation from this will very likely result in strange error
messages. The format follows the **0Highlight parts** on
[https://plantuml.com/yaml](https://plantuml.com/yaml).
}
