% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_local.R
\name{get_graph_local}
\alias{get_graph_local}
\title{Generate PlantUML graph using PlantUml jar file}
\usage{
get_graph_local(x, file = NULL)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved.
If `NULL', the graph is saved to a temporary file.}
}
\value{
name of the file with the graph
}
\description{
Generate PlantUML graph using PlantUml jar file
}
