% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_picoweb.R
\name{get_graph_picoweb}
\alias{get_graph_picoweb}
\title{Generate PlantUML graph using PicoWeb included in PlantUml jar file}
\usage{
get_graph_picoweb(x, file = NULL, type = NULL, quiet = TRUE, ...)
}
\arguments{
\item{x}{object of \code{plantuml} to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved.
If `NULL', the graph is saved to a temporary file.}

\item{type}{type of the generated graph. Only \code{png} and \code{svg} are supported.
Default: \code{svg}.}

\item{quiet}{If \code{TRUE}, suppress status messages from the download}

\item{...}{Not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Generate PlantUML graph using PicoWeb included in PlantUml jar file
}
\examples{
plantumlCode <- '
 @startuml
 (*) --> "First Activity"
 -->[You can put also labels] "Second Activity"
 --> (*)
 @enduml
'
\dontrun{
x <- as.plantuml( plantumlCode )
fn <- get_graph_server( x )
browse(fn)
}
}
