% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_server.R
\name{get_graph_server}
\alias{get_graph_server}
\title{Title}
\usage{
get_graph_server(x, file = NULL, force_png = FALSE, ...)
}
\arguments{
\item{x}{object of \code{plantuml} to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved. The extension has priority over the parameter \code{force_png}, i.e.
if the filename has the extension \code{.svg}, the parameter 'force_png\verb{is ignored. If}NULL', the graph is saved to a temporary file.}

\item{force_png}{if \code{TRUE}, a png will be returned from the plantuml server,
if \code{FALSE}, a svg will be returned and}

\item{...}{for compatibility with \code{get_graph_local()}. Not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Title
}
\examples{
plantumlCode <- '
 @startuml
 (*) --> "First Activity"
 -->[You can put also labels] "Second Activity"
 --> (*)
 @enduml
'
\dontrun{
x <- as.plantuml( plantumlCode )
fn <- get_graph_server( x )
browse(fn)
}
}
