% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getPlantumlOption}
\alias{getPlantumlOption}
\alias{setPlantumlOption}
\title{Getter and Setter for options}
\usage{
getPlantumlOption(name = NULL)

setPlantumlOption(...)
}
\arguments{
\item{name}{name of the option. if \code{NULL}, all will be returned as a \code{list.}}

\item{...}{named value to be set in the form of \code{name = value}}
}
\value{
\itemize{
\item \code{getPlantumlOption()}: the value of the option
\item \code{setPlantumlOption()}: the old value of the plantuml options
}
}
\description{
Getter and Setter for options
}
\examples{
getPlantumlOption("jar_path")

setPlantumlOption(jar_path = "something useless!")

getPlantumlOption("jar_path")

}
