% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml-package.R
\docType{package}
\name{plantuml-package}
\alias{plantuml-package}
\alias{_PACKAGE}
\title{plantuml: Create UML Graphs using PlantUML Syntax}
\description{
Create UML graphs using the PlantUML language. These are either shown in a graphics device or saved as a file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkrug/plantuml}
  \item Report bugs at \url{https://github.com/rkrug/plantuml/issues}
}

}
\author{
\strong{Maintainer}: Rainer M. Krug \email{Rainer@krugs.de} (\href{https://orcid.org/0000-0002-7490-0066}{ORCID})

}
\keyword{internal}
