% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml.R
\name{plantuml}
\alias{plantuml}
\title{Convert a character to a \code{plantuml} object}
\usage{
plantuml(x = NULL)
}
\arguments{
\item{x}{character sting containing plantuml code.}
}
\value{
object of class \code{plantuml} which can be plotted.
}
\description{
Convert a \code{character} to a \code{plantuml} object.
This can be plotted.
}
\examples{
\dontrun{
x <- '
(*) --> "Initialization"

if "Some Test" then
-->[true] "Some Activity"
--> "Another activity"
-right-> (*)
else
  ->[false] "Something else"
-->[Ending process] (*)
endif
'
x <- plantuml( x )
plot( x )
}

}
