% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_file.R
\name{plantuml_csv}
\alias{plantuml_csv}
\title{Convert structure of \code{csv} file to uml graph by loading the data}
\usage{
plantuml_csv(file, ...)
}
\arguments{
\item{file}{file name of the \code{csv} file. The function does \strong{not} do any
checking if the file is a \code{csv} file!}

\item{...}{additional arguments. Will be passed to \code{read.csv()}}
}
\value{
a \code{plantuml} containing the structure of the \code{csv} file for plotting
}
\description{
This is a convenience function which only reads the \code{csv} file using
\code{read.csv(file, ...)} and converts the resulting \code{data.frame} to a plantuml
object using \code{plantuml()}.
}
\examples{
## some preparations
x <- data.frame(
  a = c("a", "b", "c"),
  B = c("C", "D", "E"),
  x = c(0.77, 0.38, 4.43),
  bool = c(TRUE, FALSE, FALSE)
)

fn <- tempfile(fileext = ".csv")
write.csv(x, fn)

## and now the example

plot(plantuml_csv(fn))

# and only the fors columns (the rownames will not all be there)

plot(plantuml_csv(fn, nrows = 1))

# or character vectors as factors

plot(plantuml_csv(fn, as.is = FALSE))

## and cleanupo

unlink(fn)

}
