% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_file.R
\name{plantuml_json}
\alias{plantuml_json}
\title{Convert json file to uml graph}
\usage{
plantuml_json(file, text, preamble = "", ...)
}
\arguments{
\item{file}{json text to be converted. The function does \strong{not} do any
checking if the file is valid json!}

\item{preamble}{text to be inserted after the \verb{@startyaml} and before the
actual yaml. For example style and highlight info. See
\url{https://plantuml.com/yaml} for further info.}

\item{...}{additional arguments. Will be passed to \code{readLines()}}
}
\value{
a \code{plantuml} containing the json file for plotting
}
\description{
Exactly one of the two arguments (\code{file} or \code{text}) needs tobe specified.
}
\examples{
## some preparations
x <- '{"name":["Test json"],"a":["a","b","c"],"B":["C","D","E"],"x":[0.6464,0.6879]}'
fn <- tempfile(fileext = ".json")
writeLines(x, fn)

## and now the example

plot(plantuml_json(fn))

## and cleanupo

unlink(fn)

}
