% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_knit_engine.R
\name{plantuml_knit_engine}
\alias{plantuml_knit_engine}
\title{Knit engine for plantuml}
\usage{
plantuml_knit_engine(options)
}
\arguments{
\item{options}{knitr options see \code{knitr}}
}
\value{
See \code{knit_engines} in the knitr packages for details
}
\description{
This function provides a knit engine for the plantuml. It has the following additional functions:
\itemize{
\item \strong{plantuml.format}: the format of the generated image.
For formats which return images, these will be inserted,
formats resulting in text qill be included as text.
At the moment, the following are supported:
\itemize{
\item \strong{auto} (default) uses \code{svg} if output is \code{html} and \code{pdf} if output is \code{pdf}
\item \strong{png}	 To generate image using PNG format (default).
\item \strong{svg}	 To generate image using SVG format.
\item \strong{eps}	 To generate text in EPS format; or generates an image when
outputting LaTeX rather than HTML formats.
}
\item \strong{plantuml.path}: the path where the resulting files will be saved.
Default is the same directory as the \code{.Rmd} file is in in a directory named
` plantuml. The path will be created if it does not exist. The name of each
plantuml figure is the label of thre chunk with the extension as specified.
\item \strong{plantuml.preview}: if \code{TRUE}, an inline preview will be shown in RStudio.
\strong{Attention: the processing takse twice as long as without this option!}.
}
}
\details{
Thanks to Emiliano Heyns (retorquere) for giving me the idea for this
functiion https://github.com/rkrug/plantuml/issues/10#issue-639795529
See \code{knit_engines} in the knitr packages for details
}
\examples{
\dontrun{
## see
system.file("plantuml.Rmd", package = "plantuml")
## for an example RMarkdown file usint this knit engine.
}
}
