% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_file.R
\name{plantuml_rds}
\alias{plantuml_rds}
\title{Convert structure of \code{rds} file to uml graph by loading the data}
\usage{
plantuml_rds(file, ...)
}
\arguments{
\item{file}{file name of the \code{rds} file. The function does \strong{not} do any
checking if the file is a \code{rds} file!}

\item{...}{additional arguments. Will be passed to \code{readRDS()}}
}
\value{
a \code{plantuml} containing the structure of the \code{rds} file for plotting
}
\description{
This is a convenience function which only reads the \code{rds} file using
\code{readRDS(file, ...)} and converts the resulting object to a plantuml
object using \code{plantuml()}.
}
\examples{
## some preparations
x <- list(
  name = "Test list",
  a = c("a", "b", "c"),
  B = c("C", "D", "E"),
  x = c(0.776318477466702, 0.381654617609456)
)

fn <- tempfile(fileext = ".rds")
saveRDS(x, fn)

## and now the example

plot(plantuml_rds(fn))

## and cleanupo

unlink(fn)

}
