% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodePlantUMLServer.R
\name{plantuml_URL}
\alias{plantuml_URL}
\title{Generate PlantUML Server URL}
\usage{
plantuml_URL(
  plantuml = "@startuml\\nBob -> Alice : hello\\n@enduml",
  server_url = "http://www.plantuml.com/plantuml/png/",
  open_in_browser = FALSE
)
}
\arguments{
\item{plantuml}{The plantuml code}

\item{server_url}{The base url of the server}
}
\value{
complete url
}
\description{
Generate PlantUML Server URL
}
\examples{

}
