% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml.R
\name{plotPlantuml}
\alias{plotPlantuml}
\title{Generate UML graph from plantuml}
\usage{
plotPlantuml(plantuml, file = NULL, plantumlOpt = "")
}
\arguments{
\item{plantuml}{plantuml code to draw the UML graph}

\item{file}{if \codew{file} is \code{NULL}, a png is created, saved in a
temporary file and drawn in a device. If file is a file name, the graph is
saved in the file and the type is based on the extensions. See limitations
of plantuml to get the list of available file formats]}

\item{plantumlOpt}{additional options for plantuml in addition to \code{-p} and \code{-tFILETYPE}}
}
\value{
returns file name (including absolute path) of the created graph.
}
\description{
Generate an image containing based ion the plantuml code.
TODO use pipe instead of temporary file
TODO scale window according to png
TODO can I use vector formats?
}
