% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server.R
\name{server_available}
\alias{server_available}
\title{Check if server is available}
\usage{
server_available()
}
\value{
\code{TRUE} if server is available, \code{FALSE} otherwise
}
\description{
This function tries to download a file from the server.
This is suboptimal - suggestions welcome.
}
