% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_info.R
\name{server_info}
\alias{server_info}
\title{Get info about the plantuml server}
\usage{
server_info()
}
\value{
list with the following arguments:
\itemize{
\item \strong{server}: type of the server, \strong{PicoWeb Server} or \strong{PlantUML Web Server}
\item **supported_formats: supported formats of the server
}
}
\description{
The parameter (URL and port) as in the \code{getPlantumlOption()} specified are used.
}
