% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_set.R
\name{server_set}
\alias{server_set}
\title{Set parameter of the plantuml server addrss to the default valus}
\usage{
server_set(location = "remote")
}
\arguments{
\item{location}{alloeed values are:
\itemize{
\item "remote": default values for the remote server
\item "local": default values for a local picoweb server. The serer still needs to be started!
Other locations need to be set manually by setting the options directly.
}}
}
\value{
invisibly the old values of plantuml options
}
\description{
Set parameter of the plantuml server addrss to the default valus
}
