% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_start.R
\name{server_start}
\alias{server_start}
\title{Start a PlantUML Picoweb Server on localhost}
\usage{
server_start()
}
\value{
\code{integer} or \code{list()} as returned from \code{server_status()}
\itemize{
\item \strong{-999}: unknown error
\item \strong{1}  : unknown server running on the port on localhost
\item \strong{2}   : PlantUML Web Server running on that port on localhost
\item \strong{\code{server_info()}} ehen PicoWeb Server already running or started
}
}
\description{
If the port \code{getPlantumlOption("server_port")} on localhostis is available,
i.e. no other server is running on that port, a PantUML PicoWeb Server is started.
The return value indicates if a server was started and if not, why.
}
