% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_status.R
\name{server_status}
\alias{server_status}
\title{Get the status of the server}
\usage{
server_status()
}
\value{
\code{integer} value indicating the status:
\itemize{
\item \strong{-999}: undetermined error (should not happen)!
\item \strong{-404}: URL not reachable
\item \strong{0}  : server reachable but not a known PlantUml Web Server or PicWeb Plantuml Server
\item \strong{1}  : server is PicoWeb Server
\item \strong{2}  : server is PlantUML Web Server
}
}
\description{
The parameter (URL and port) as in the \code{getPlantumlOption()} specified are used.
}
