% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_stop.R
\name{server_stop}
\alias{server_stop}
\title{Stop a PlantUML Picoweb Server running on localhost}
\usage{
server_stop()
}
\value{
\code{integer} or \code{list()} as returned from \code{server_status()}
\itemize{
\item \strong{-999}: unknown error
\item \strong{-1}  : unknown server running on the port on localhost
\item \strong{-2}  : PlantUML Web Server running on that port on localhost
\item \strong{0}   : PicoWeb Server in the process of stopping or no server running on tat port
}
}
\description{
The server is expscted at the port \code{getPlantumlOption("server_port")} on localhost.
}
