% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server.R
\name{set_method}
\alias{set_method}
\title{Set the method to be used to generate the plantUML graph}
\usage{
set_method(method = "auto", check_server = TRUE)
}
\arguments{
\item{method}{Method to be used to generate the plantUML graph
\itemize{
\item \code{auto}: irrespective of the value of \code{check_server}, the method
is set to \code{local} if the server is not available
(\code{server_available() == FALSE}), and to \code{server}
if it is available (\code{server_available() == TRUE})
}}

\item{check_server}{If \code{TRUE}, the server availability is checked
before setting to \code{server} and if it does not exist, set to \code{local}}
}
\value{
the new value of \code{method}
}
\description{
The plantuml option \code{method} is set. When called with default values, it
will set the method to \code{server} if the server can be reached, otherwise to \code{local}
}
