% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.plantuml.R
\name{as.plantuml}
\alias{as.plantuml}
\title{Convert an R object to a Class uml object}
\usage{
as.plantuml(x, ...)
}
\arguments{
\item{x}{R object to be converted}

\item{...}{other argunents passed on to generic functions}
}
\value{
object of class \code{plantuml} which can be plotted.
}
\description{
Generic function to convert an R object to a \code{plantuml} Class object.
This can be plotted.
}
\examples{
\dontrun{
x <- list(
  a = 1:10,
  b = letters[1:4],
  c = data.frame(
    x = 1:10,
    y = c(TRUE, FALSE)
  )
)
x <- plantuml(x)
plot( x )
}

}
