% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.plantuml.R
\name{as.plantuml}
\alias{as.plantuml}
\title{Convert a character to a \code{plantuml} object}
\usage{
as.plantuml(x)
}
\arguments{
\item{x}{plantuml code as a character string.}
}
\value{
object of class \code{plantuml} which can be plotted.
}
\description{
Convert a \code{character} to a \code{plantuml} object. This can be plotted.
}
\examples{
x <- '
 @startuml
  --> "First Activity"
 -->[You can put also labels] "Second Activity"
 -->
 @enduml
'
x <- as.plantuml( x )
plot( x )
}
