% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.plantuml.R, R/as.plantuml.character.R,
%   R/as.plantuml.complex.R, R/as.plantuml.data.frame.R,
%   R/as.plantuml.default.R, R/as.plantuml.factor.R, R/as.plantuml.integer.R,
%   R/as.plantuml.list.R, R/as.plantuml.logical.R, R/as.plantuml.numeric.R
\name{as.plantuml}
\alias{as.plantuml}
\alias{as.plantuml.character}
\alias{as.plantuml.complex}
\alias{as.plantuml.data.frame}
\alias{as.plantuml.default}
\alias{as.plantuml.factor}
\alias{as.plantuml.integer}
\alias{as.plantuml.list}
\alias{as.plantuml.logical}
\alias{as.plantuml.numeric}
\title{Convert an R object to a Class uml object}
\usage{
as.plantuml(x, ...)

\method{as.plantuml}{character}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{complex}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{data.frame}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{default}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{factor}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{integer}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{list}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{logical}(x, complete = FALSE, nm = NULL, ...)

\method{as.plantuml}{numeric}(x, complete = FALSE, nm = NULL, ...)
}
\arguments{
\item{x}{R object to be converted}

\item{...}{other argunents passed on to generic functions}
}
\value{
object of class \code{plantuml} which can be plotted.
}
\description{
Generic function to convert an R object to a \code{plantuml} Class object.
See details for the individual functions
}
\examples{
\dontrun{
x <- list(
  a = 1:10,
  b = letters[1:4],
  c = data.frame(
    x = 1:10,
    y = c(TRUE, FALSE)
  )
)
x <- plantuml(x)
plot( x )
}

\dontrun{
 ## \code{character}
 x <- LETTERS
 x <- as.plantuml(x)
 plot(x)
 }

\dontrun{
 ## \code{complex}
 x <- 1:2 + 1i*(8:9)
 x <- as.plantuml(x)
 plot(x)
 }

\dontrun{
 ## \code{factor}
 x <- factor(LETTERS)
 x <- as.plantuml(x)
 plot(x)
 }

\dontrun{
 ## \code{integer}
 x <- 1L:10L
 names(x) <- letters[1:10]
 x <- as.plantuml(x)
 plot(x)
 }

\dontrun{
 ## \code{list}
 x <- list(
   num = 1:10,
   let = letters[1:20]
 )
 x$list <- list(x, x)
 x <- as.plantuml(x)
 plot(x)
 }

\dontrun{
  ## \code{logical}
  x <- c(True, FALSE, TRUE, NA)
  x <- as.plantuml(x)
  plot(x)
 }

\dontrun{
 ## \code{numeric}
 x <- 1/(-10:10)
 x <- as.plantuml(x)
 plot(x)

 x <- matrix(1:21, nrow = 7, ncol = 3)
 x <- as.plantuml(x)
 plot(x)
 }

}
