% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph.R
\name{get_graph}
\alias{get_graph}
\title{Title}
\usage{
get_graph(x, file = NULL, width = NULL, height = NULL, css = NULL, ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved. The extension has priority over the parameter \code{force_png}, i.e.
if the filename has the extension \code{.svg}, the parameter 'force_png\verb{is ignored. If}NULL', the graph is saved to a temporary file.}

\item{width}{output width in pixels or NULL for default.}

\item{height}{output height in pixels or NULL for default}

\item{css}{path/url to external css file or raw vector with css data.
This requires your system has a recent version of librsvg.}

\item{...}{Additional values for the function
\code{get_graph_local()} or \code{get_graph_server()}
but not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Title
}
