% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph.R
\name{get_graph}
\alias{get_graph}
\title{Title}
\usage{
get_graph(x, file = NULL, vector = TRUE, ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{\itemize{
\item \strong{\code{file} is `NULL}: the graph is drawn in the graphics device.
\item \strong{\code{file} is a file name}: the graph is saved in the file and the type
is based on the extensions.
}}

\item{vector}{if \code{TRUE} use \code{svg} as intermediate format, if \code{FALSE}
use \code{png}. Only used when file is \code{NULL}.}

\item{...}{Additional values for the function \code{get_graph_local()}.}
}
\value{
name of the file with the graph
}
\description{
Title
}
