% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_local.R
\name{get_graph_local}
\alias{get_graph_local}
\title{Title}
\usage{
get_graph_local(x, file = NULL, width = NULL, height = NULL, css = NULL, ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved.
If `NULL', the graph is saved to a temporary file.}

\item{width}{output width in pixels or NULL for default.}

\item{height}{output height in pixels or NULL for default}

\item{css}{path/url to external css file or raw vector with css data.
This requires your system has a recent version of librsvg.}

\item{...}{Not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Title
}
