% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_local.R
\name{get_graph_local}
\alias{get_graph_local}
\title{Title}
\usage{
get_graph_local(x, file = NULL, force_png = FALSE, ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved. The extension has priority over the parameter \code{force_png}, i.e.
if the filename has the extension \code{.svg}, the parameter 'force_png\verb{is ignored. If}NULL', the graph is saved to a temporary file.}

\item{...}{for compatibility with \code{get_graph_server()}. Not used at the moment.}

\item{plantuml_opt}{additional options for plantuml in addition to \code{-p}
and \code{-tFILETYPE}. See `plantuml_run() for a list of available file formats.}
}
\value{
name of the file with the graph
}
\description{
Title
}
