% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_local.R
\name{get_graph_local}
\alias{get_graph_local}
\title{Title}
\usage{
get_graph_local(x, file = NULL, vector = TRUE, plantuml_opt = "", ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{\itemize{
\item \strong{\code{file} is `NULL}: the graph is drawn in the graphics device.
\item \strong{\code{file} is a file name}: the graph is saved in the file and the type
is based on the extensions.
}}

\item{vector}{if \code{TRUE} use \code{svg} as intermediate format, if \code{FALSE}
use \code{png}. Only used when file is \code{NULL}.}

\item{plantuml_opt}{additional options for plantuml in addition to \code{-p}
and \code{-tFILETYPE}. See `plantuml_run() for a list of available file formats.}

\item{...}{for compatibility with \code{get_graph_server()}. Not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Title
}
