% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_server.R
\name{get_graph_server}
\alias{get_graph_server}
\title{Title}
\usage{
get_graph_server(
  x,
  file = NULL,
  width = NULL,
  height = NULL,
  css = NULL,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{x}{object of \code{plantuml} to draw the UML graph}

\item{file}{file name, including extension, to which the returned plantUML graph
should be saved.
If `NULL', the graph is saved to a temporary file.}

\item{width}{output width in pixels or NULL for default.}

\item{height}{output height in pixels or NULL for default}

\item{css}{path/url to external css file or raw vector with css data.
This requires your system has a recent version of librsvg.}

\item{quiet}{If \code{TRUE}, suppress status messages from the download}

\item{...}{Not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Title
}
\examples{
plantumlCode <- '
 @startuml
 (*) --> "First Activity"
 -->[You can put also labels] "Second Activity"
 --> (*)
 @enduml
'
\dontrun{
x <- as.plantuml( plantumlCode )
fn <- get_graph_server( x )
browse(fn)
}
}
