% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_server.R
\name{get_graph_server}
\alias{get_graph_server}
\title{Title}
\usage{
get_graph_server(x, file = NULL, vector = TRUE, ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{\itemize{
\item \strong{\code{file} is `NULL}: the graph is drawn in the graphics device.
\item \strong{\code{file} is a file name}: the graph is saved in the file and the type
is based on the extensions.
}}

\item{vector}{if \code{TRUE} use \code{svg} as intermediate format, if \code{FALSE}
use \code{png}. Only used when file is \code{NULL}.}

\item{...}{for compatibility with \code{get_graph_local()}. Not used at the moment.}
}
\value{
name of the file with the graph
}
\description{
Title
}
\examples{
plantumlCode <- '
 @startuml
 (*) --> "First Activity"
 -->[You can put also labels] "Second Activity"
 --> (*)
 @enduml
'
\dontrun{
x <- as.plantuml( plantumlCode )
get_graph_server( x )
plot(as.plantuml(x), java_opt = "-Djava.awt.headless=true")
}
}
