% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getPlantumlOption}
\alias{getPlantumlOption}
\alias{plantumlOptions}
\title{Getter and Setter for options}
\usage{
getPlantumlOption(x, default = NULL)

plantumlOptions(...)
}
\arguments{
\item{x}{a character string holding an option name.}

\item{default}{if the specified option is not set in the options list, this
value is returned. This facilitates retrieving an option and checking
whether it is set and setting it separately if not.}

\item{...}{named value to be set in the form of \code{name = value}}

\item{plantuml_opt}{additional options for plantuml in addition to \code{-p}
and \code{-tFILETYPE}. See `plantuml_run() for a list of available file formats.}

\item{vector}{if \code{TRUE} use \code{svg} as intermediate format, if \code{FALSE}
use \code{png}. Only effects plotting in device.}
}
\value{
\itemize{
\item \code{getPlantumlOption()}: the value of the option
\item \code{plantumlOptions()}: the old value of the plantuml options
}
}
\description{
Allow the user to set and examine the parameter used by the package \code{plantuml}
}
\examples{
getPlantumlOption("jar_name")

plantumlOptions(jar_name = "something useless!")

getPlantumlOption("jar_name")

}
