% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_file.R
\name{plantuml_file}
\alias{plantuml_file}
\title{Convert file to plantuml code}
\usage{
plantuml_file(file, preamble = "", ...)
}
\arguments{
\item{file}{file name. At the moment the following files are supported:
\itemize{
\item \code{.yml} or \code{yaml}: the yml data is graphed
\item \code{.json}: the json file is graphed
\item \code{.rds} : the file is loaded and the object contained is graphed
\item \code{.csv} : the csv is loaded and graphed
}}

\item{preamble}{\strong{Only used for yaml and json files.} Text to be inserted
after the \verb{@startyaml} and before the actual yaml. For example style and
highlight info. See \url{https://plantuml.com/yaml}
for further info.}
}
\description{
The function either plots the object including values (\code{yaml} and \code{json}) or
loads the object and plots the structure (\code{csv} and \code{rds}).
}
\examples{
\dontrun{
plantuml_file("name.yml")
plantuml_file("name.yaml")
plantuml_file("name.json")
plantuml_file("name.rds")
plantuml_file("name.csv")
}
}
