% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_file.R
\name{plantuml_json}
\alias{plantuml_json}
\title{Convert json file to uml graph}
\usage{
plantuml_json(file, ...)
}
\arguments{
\item{file}{file name of the json file. The function does \strong{not} do any
checking if the file is a json file!}

\item{...}{additional arguments. Will be passed to \code{readLines()}}
}
\value{
a \code{plantuml} containing the json file for plotting
}
\description{
Convert json file to uml graph
}
\examples{
## some preparations
x <- '{"name":["Test json"],"a":["a","b","c"],"B":["C","D","E"],"x":[0.6464,0.6879]}'
fn <- tempfile(fileext = ".json")
writeLines(x, fn)

## and now the example

plot(plantuml_json(fn))

## and cleanupo

unlink(fn)

}
