% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_update.R
\name{plantuml_update}
\alias{plantuml_update}
\title{Update or download \code{plantuml.jar} binary}
\usage{
plantuml_update(tag = "release")
}
\arguments{
\item{tag}{tag of the version to be downloaded. Allowed values are
\itemize{
\item \strong{"release}    : the last release
\item \strong{"snapshot"}  : the last snapshot - \strong{Not ready for gneral use}
}}
}
\value{
the path and name of the downloaded file
}
\description{
The file \code{plantuml.jar} is downloaded from Sourceforge and saved as
\code{"plantuml.jar"} in the folder \code{system.file("jar", package =
"plantuml")} of the package. The source code for \code{plantuml} can be
found at \url{https://github.com/plantuml/plantuml}
}
\examples{
\dontrun{
plantuml_update()
}
}
