% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantUML_URL.R
\name{plantuml_URL}
\alias{plantuml_URL}
\title{Generate PlantUML Server URL}
\usage{
plantuml_URL(
  plantuml = "@startuml\\nBob -> Alice : hello\\n@enduml",
  server_url = getPlantumlOption("server_url"),
  server_port = getPlantumlOption("server_port"),
  type = "svg",
  open_in_browser = FALSE
)
}
\arguments{
\item{plantuml}{The plantuml code.
If \code{NULL}, only the base URL consisting of server and (if specified) port will be returned.}

\item{server_url}{The base url of the server. Should end with a single \code{/}
If \code{NULL}, encoding, including the ttpe, will be returned.}

\item{server_port}{port on which the plantuml server is}

\item{type}{the type of the returned figure. At the moment supported:
\code{png}, \code{svg} \code{txt}, or \code{map}. See \url{https://plantuml.com/server} for further details.
If \code{NULL}, encoding, excluding the tytpe, will be returned.
\strong{The return value is invalid if \code{server_url} is given!}}

\item{open_in_browser}{if TRUE, the result will be opened in the browser.}
}
\value{
complete url to retrieve the graph.
}
\description{
Generate PlantUML Server URL
}
