% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantUML_URL.R
\name{plantuml_URL}
\alias{plantuml_URL}
\title{Generate PlantUML Server URL}
\usage{
plantuml_URL(
  plantuml = "@startuml\\nBob -> Alice : hello\\n@enduml",
  server_url = getPlantumlOption("server_url"),
  type = "png",
  open_in_browser = FALSE
)
}
\arguments{
\item{plantuml}{The plantuml code}

\item{server_url}{The base url of the server. Should end with a single \code{/}}

\item{type}{the type of the returned figure. At the moment supported:
\code{png}, \code{svg} or \code{txt} for ASCIIArt.}

\item{open_in_browser}{if TRUE, the result will be opened in the browser.}
}
\value{
complete url to retrieve the graph.
}
\description{
Generate PlantUML Server URL
}
\examples{

}
