% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_file.R
\name{plantuml_yaml}
\alias{plantuml_yaml}
\title{Convert yaml file to uml graph}
\usage{
plantuml_yaml(file, ...)
}
\arguments{
\item{file}{file name of the yaml file. The function does \strong{not} do any
checking if the file is a yaml file!}

\item{...}{additional arguments. Will be passed to \code{readLines()}}
}
\value{
a \code{plantuml} containing the yaml file for plotting
}
\description{
Convert yaml file to uml graph
}
\examples{
## some preparations
x <- "name: Test yml\na:\n- a\n- b\n- c\nB:\n- C\n- D\n- E\nx:\n- 0.2486851\n- 0.7498182\n"
fn <- tempfile(fileext = ".yml")
writeLines(x, fn)

## and now the example

plot(plantuml_yaml(fn))

## and cleanupo
unlink(fn)

}
