% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.plantuml.R
\name{plot.plantuml}
\alias{plot.plantuml}
\title{Generate UML graph from plantuml}
\usage{
\method{plot}{plantuml}(x, file = NULL, plantuml_opt = "",
  vector = TRUE, java_opt = "", ...)
}
\arguments{
\item{x}{plantuml code to draw the UML graph}

\item{file}{if \code{file} is \code{NULL}, a png is created, saved in a
temporary file and drawn in a device. If file is a file name, the graph is
saved in the file and the type is based on the extensions. See limitations
of plantuml to get the list of available file formats]}

\item{plantuml_opt}{additional options for plantuml in addition to \code{-p} and \code{-tFILETYPE}}

\item{vector}{if \code{TRUE} use svg as intermediate format, if \code{FALSE} use png. Only effects plotting in device.}

\item{java_opt}{additional options for java. see java documentation for details.}

\item{...}{additional arguments for the plot function \code{grid::grid.raster()}}
}
\value{
returns file name (including absolute path) of the created graph.
}
\description{
Generate an image containing based ion the plantuml code.
TODO can I use vector formats?
}
\examples{
plantumlCode <- '
 @startuml
 (*) --> "First Activity"
 -->[You can put also labels] "Second Activity"
 --> (*)
 @enduml
'
\dontrun{
x <- as.plantuml( plantumlCode )
plot( x )
plot(as.plantuml(x), java_opt = "-Djava.awt.headless=true")
}
}
