% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePlantumlJar.R
\name{updatePlantumlJar}
\alias{updatePlantumlJar}
\title{Update or download \code{plantuml.jar} binary}
\usage{
updatePlantumlJar(beta = FALSE, ...)
}
\arguments{
\item{beta}{should the beta version be downloaded, default \code{FALSE}}

\item{...}{additional arguments for the \code{download.file()} function}
}
\value{
the result from the \code{download.file()}
}
\description{
The file \code{plantuml.jar} is downloaded from Sourceforge and saved as
\code{"plantuml.jar"} in the folder \code{system.file("jar", package =
"plantuml")} of the package.
}
\examples{
\dontrun{
updatePlantumlJar()
}
}
