% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuml_update.R
\name{updatePlantumlJar}
\alias{updatePlantumlJar}
\title{RENAMED to plantuml_update()!}
\usage{
updatePlantumlJar(...)
}
\arguments{
\item{...}{some parameter}
}
\description{
RENAMED to plantuml_update()!
}
